backend/requirements.txt
fastapi
uvicorn
pillow
pyzbar
python-multipart